# HackathonServer/src/main/java/com/example/HackathonServer/models/Score.java
package com.example.HackathonServer.models;

import com.fasterxml.jackson.annotation.JsonBackReference;
import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Entity
@Data
@NoArgsConstructor
@AllArgsConstructor
public class Score {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

//    @ManyToOne(fetch = FetchType.LAZY)
//    @JoinColumn(name = "child_id", nullable = false)
//    private Child child;

//    @ManyToOne(fetch = FetchType.LAZY)
//    @JoinColumn(name = "video_id", nullable = false)
//    private Video video;

    @Column(nullable = false)
    private Integer score;

    @Column(nullable = false)
    private Double percentage;
    @OneToOne(fetch = FetchType.LAZY)
    @JsonBackReference(value = "session-score")
    private Session session;
}