# project/docs/apis/exergame-openapi.yaml
openapi: 3.0.3
info:
  title: Exergame API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /move:
    get:
      tags:
        - Move
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Move"
        "400":
          description: Bad Request
        "404":
          description: Not Found
    post:
      tags:
        - Move
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Move"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Move"
  /move/{id}:
    get:
      tags:
        - Move
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Move"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
    put:
      tags:
        - Move
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Move"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Move"
    delete:
      tags:
        - Move
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
  /move/last7days/{childId}:
    get:
      tags:
        - Move
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: integer
            format: int64
  /move/last30days/{childId}:
    get:
      tags:
        - Move
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: integer
            format: int64
...
  /score:
    get:
      tags:
        - Score
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"
        "400":
          description: Bad Request
        "404":
          description: Not Found
    post:
      tags:
        - Score
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Score"
  /score/{id}:
    get:
      tags:
        - Score
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
    put:
      tags:
        - Score
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Score"
    delete:
      tags:
        - Score
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
  /score/{childId}/last-week/average:
    get:
      tags:
        - Score
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: number
                format: double
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: integer
            format: int64
  /score/{childId}/last-week/grade:
    get:
      tags:
        - Score
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: integer
            format: int64
  /score/{childId}/last-month/average:
    get:
      tags:
        - Score
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: number
                format: double
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: integer
            format: int64
  /score/{childId}/last-month/grade:
    get:
      tags:
        - Score
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: integer
            format: int64
...
components:
  schemas:
...
    Move:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Score:
      type: object
      properties:
        id:
          type: integer
          format: int64
        child:
          type: string
        video:
          type: string
        score:
          type: integer
          format: int32
        percentage:
          type: number
          format: double
      required:
        - percentage
        - score
...